+++ eslint.config.js
@@ -8,4 +8,9 @@ export default [
         ignores: ['build/**'],
     },
     ...eslintConfigTs,
+    {
+        rules: {
+            '@typescript-eslint/consistent-type-imports': 'error',
+        },
+    },
 ];